chore: spec work FE
