+++ bciers/apps/administration/app/components/operations/OperationInformationForm.tsx
@@ -21,7 +21,7 @@ import { FormMode, FrontEndRoles } from "@bciers/utils/src/enums";
 import { useSessionRole } from "@bciers/utils/src/sessionUtils";
 import Note from "@bciers/components/layout/Note";
 import Link from "next/link";
-import ConfirmChangeOfRegistrationPurposeModal from "@/registration/app/components/operations/registration/ConfirmChangeOfRegistrationPurposeModal";
+import ConfirmChangeOfFieldModal from "@/registration/app/components/operations/registration/ConfirmChangeOfFieldModal";
 
 const OperationInformationForm = ({
   formData,
@@ -137,10 +137,12 @@ const OperationInformationForm = ({
           contact to replace them.
         </Note>
       )}
-      <ConfirmChangeOfRegistrationPurposeModal
+      <ConfirmChangeOfFieldModal
         open={isConfirmPurposeChangeModalOpen}
-        cancelRegistrationPurposeChange={cancelRegistrationPurposeChange}
-        confirmRegistrationPurposeChange={confirmRegistrationPurposeChange}
+        onCancel={cancelRegistrationPurposeChange}
+        onConfirm={confirmRegistrationPurposeChange}
+        modalText="Are you sure you want to change your registration purpose? If you proceed, some of the form data you have entered will be lost."
+        confirmButtonText="Change registration purpose"
       />
       <SingleStepTaskListForm
         allowEdit={!role.includes("cas_")}
